+++ src/lib.rs
@@ -27,10 +27,10 @@
 //! XChaCha20Poly1305.
 //!
 //! ## Password hashing and verification
-//! [`orion::pwhash`] offers password hashing and verification using Argon2i.
+//! [`orion::pwhash`] offers password hashing and verification using Argon2id.
 //!
 //! ## Key derivation
-//! [`orion::kdf`] offers key derivation using Argon2i.
+//! [`orion::kdf`] offers key derivation using Argon2id.
 //!
 //! ## Message authentication
 //! [`orion::auth`] offers message authentication and verification using BLAKE2b.