+++ tests/kdf/pynacl_argon2i.rs
@@ -3,7 +3,7 @@
 // The generated test vectors have been generated the 24th January 2020.
 
 use hex::decode;
-use orion::hazardous::kdf::argon2i;
+use orion::hazardous::kdf::argon2::*;
 use serde::{Deserialize, Serialize};
 use std::{fs::File, io::BufReader};
 
@@ -26,14 +26,13 @@ fn run_tests_from_json(path_to_vectors: &str) {
 
     for test in tests {
         let mut dst_out = vec![0u8; test.dgst_len];
-
+        let cost = CostParams::new(test.iters, test.maxmem, 1).unwrap();
         assert!(
-            argon2i::verify(
+            Argon2::<I, Sequential>::verify(
                 &decode(&test.pwhash).unwrap(),
                 test.passwd.as_bytes(),
                 test.salt.as_bytes(),
-                test.iters,
-                test.maxmem,
+                &cost,
                 None,
                 None,
                 &mut dst_out