+++ CHANGELOG.md
@@ -33,6 +33,36 @@
 	- [Breaking change]: Remove `chacha20poly1305::seal()`, `chacha20poly1305::open()`, `xchacha20poly1305::seal()` `xchacha20poly1305::open()`.
 	- Add structs `ChaCha20Poly1305` and `XChaCha20Poly1305` that offer equivalent `open()` and `seal()` functions from versions prior to `0.18.0`.
 	- Add support for `seal_inplace()` and `open_inplace()` for `ChaCha20Poly1305` and `XChaCha20Poly1305`. These overwrite data directly instead of copying and allow handling the authentication tag separately.
+- [Breaking change] `orion::hazardous::kdf::argon2`:
+	- Functions moved to `Argon2<Variant, Threading>` struct.
+	- Added `Argon2::derive_key_encoded()` and `Argon2::verify_encoded()` that work on P-H-C strings.
+	- Hazardous now contains `safe_api` feature-gated `PasswordHash`.
+	- Add `CostParams` struct being passes to functions.
+- [Breaking change] `orion::pwhash` now uses Argon2id:
+	- See [doc/MIGRATION](doc/MIGRATION.md) on how to migrate existing `0.17` `orion::pwhash::PasswordHash`es.
+	- `PasswordHash::unprotected_as_encoded()` -> `PasswordHash::unprotected_as_str()`.
+	- `PasswordHash` is nonw a wrapper around the PHC-encoded string only:
+		- `PasswordHash::from_slice()` is removed. `TryFrom<&[u8]>` exists, but expects this to be byte-repr of a PHC-encoded string. There is no longer
+		and option to make a `PasswordHash` manually from byte-repr hash annd the parameters.
+		- `PasswordHash::len()` returns the length of the encoded string.
+	- Passes cost parameters with new `CostParams` struct.
+	- Restriction of minimum iterations of `3` is removed as this does not apply to Argon2id.
+	- Parallelism/lane cost parameter added which scales minimum memory requirement.
+- [Breaking change] `orion::kdf` now uses Argon2id:
+	- See [doc/MIGRATION](doc/MIGRATION.md) on how to migrate existing `0.17` `orion::kdf`es.
+	- Passes cost parameters with new `CostParams` struct.
+	- Restriction of minimum iterations of `3` is removed as this does not apply to Argon2id.
+	- Parallelism/lane cost parameter added which scales minimum memory requirement.
+- [Breaking change] `orion::hazardous::kdf::scrypt`:
+	- Functions moved to `Scrypt` struct.
+	- Added `Scrypt::derive_key_encoded()` and `Scrypt::verify_encoded()` that work on P-H-C strings.
+	- Hazardous now contains `safe_api` feature-gated `PasswordHash`.
+	- Add `CostParams` struct being passes to functions.
+	- Cost parameter `n` is now `logn` to make choosing wrong combinations harder.
+- [Breaking change] `orion::hazardous::kdf::pbkdf2`:
+	- Functions moved to `Pbkdf2` struct.
+	- `password` being passed is now a byte-slice, instead of a HMAC key that internally pads the key (yielding the padded if called `unprotected_as_ref()` on).
+- Add support for Argon2id (sequential-mode only).
 - MSRV bumped to `1.87`
 - Add constants for BLAKE2b: `BLAKE2B_MIN_OUTSIZE, BLAKE2B_MAX_OUTSIZE, BLAKE2B_MIN_KEYSIZE, BLAKE2B_MAX_KEYSIZE` making the conditions more discernable.
 - Add support for the following PQ/T HPKE suites: